x-monitoring-be/requirements.txt
# Python 3.13+
# JAVA_HOME 환경변수 필요 (JDK 11+)

# Web framework
Flask==3.1.3
flask-cors==6.0.2
flask-limiter==4.1.1

# Validation & security
pydantic==2.12.5
PyJWT==2.12.1
python-dotenv==1.2.2

# JDBC bridge (JayDeBeApi depends on JPype1)
JayDeBeApi==1.2.3
JPype1==1.6.0

# SSH remote command execution (server resource monitoring)
paramiko==4.0.0

# HTTP client (health-check proxy)
requests==2.33.1

# EXE packaging
pyinstaller==6.19.0

